requirements.txt
python-dotenv
spacy
svgwrite
TeleBot